Moved cleanupPathInfo to Routing
