+++ libraries/classes/Config.php
@@ -837,7 +837,7 @@ class Config
             }
         }
 
-        $parsedUrlPath = parse_url($this->cleanupPathInfo(), PHP_URL_PATH);
+        $parsedUrlPath = parse_url(Routing::getCleanPathInfo(), PHP_URL_PATH);
 
         $parts = explode(
             '/',
@@ -857,56 +857,6 @@ class Config
         $parts[] = '';
 
         return implode('/', $parts);
-    }
-
-    /**
-     * PATH_INFO could be compromised if set, so remove it from PHP_SELF
-     * and provide a clean PHP_SELF here
-     */
-    private function cleanupPathInfo(): string
-    {
-        $pmaPhpSelf = Core::getenv('PHP_SELF');
-        if ($pmaPhpSelf === '') {
-            $pmaPhpSelf = urldecode(Core::getenv('REQUEST_URI'));
-        }
-
-        $_PATH_INFO = Core::getenv('PATH_INFO');
-        if ($_PATH_INFO !== '' && $pmaPhpSelf !== '') {
-            $question_pos = mb_strpos($pmaPhpSelf, '?');
-            if ($question_pos != false) {
-                $pmaPhpSelf = mb_substr($pmaPhpSelf, 0, $question_pos);
-            }
-
-            $path_info_pos = mb_strrpos($pmaPhpSelf, $_PATH_INFO);
-            if ($path_info_pos !== false) {
-                $path_info_part = mb_substr($pmaPhpSelf, $path_info_pos, mb_strlen($_PATH_INFO));
-                if ($path_info_part === $_PATH_INFO) {
-                    $pmaPhpSelf = mb_substr($pmaPhpSelf, 0, $path_info_pos);
-                }
-            }
-        }
-
-        $path = [];
-        foreach (explode('/', $pmaPhpSelf) as $part) {
-            // ignore parts that have no value
-            if ($part === '' || $part === '.') {
-                continue;
-            }
-
-            if ($part !== '..') {
-                // cool, we found a new part
-                $path[] = $part;
-            } elseif ($path !== []) {
-                // going back up? sure
-                array_pop($path);
-            }
-
-            // Here we intentionall ignore case where we go too up
-            // as there is nothing sane to do
-        }
-
-        /** TODO: Do we really need htmlspecialchars here? */
-        return htmlspecialchars('/' . implode('/', $path));
     }
 
     /**
